TT-619: URLs - update service start page URL
Move Service start page URL question up to the top of the URLs (as it's
the most obvious one). Update the copy to match the form here
https://docs.google.com/a/digital.cabinet-office.gov.uk/document/d/1HvW0dC07OFH8HraaS-6b7XQSkzql7Y_weg_VTFQRY1o/edit?usp=sharing

As usual, I've left the "More Information" toggle in place.

I'm not sure about the "(Refer: ...)" language. I would have gone with
"(See the [Service Manual's guidance on start pages])".

Authors: @richardtowers