TT-619: Update copy for matching service URL
In line with
https://docs.google.com/a/digital.cabinet-office.gov.uk/document/d/1HvW0dC07OFH8HraaS-6b7XQSkzql7Y_weg_VTFQRY1o/edit?usp=sharing

I haven't switched the URL formats based on environment yet - I'll look
at doing this in a later commit.

I've deviated from the copy slightly because it mentioned "range of
URLs" which is not something people can really provide. I'm also not
sure that this is the right place to be introducing the idea that we
need to whitelist IPs.

Authors: @richardtowers