TT-619: Update copy for Test device browser IP addresses
Mostly as per
https://docs.google.com/a/digital.cabinet-office.gov.uk/document/d/1HvW0dC07OFH8HraaS-6b7XQSkzql7Y_weg_VTFQRY1o/edit?usp=sharing
.

I've executively gone with Jennifer's suggestion of a full stop rather
than a colon. Feels better as two sentences to me, but happy to change
if people disagree.

I've also removed the quotes from around the addresses as they're not
really necessary when styled as <code>.

Authors: @richardtowers