Zendesk group ID as environment variable and doc improvements - Take 3
This puts back in the changes originally done in https://github.com/alphagov/verify-rp-environment-config-form/pull/29

Original commit message:

The PR brings the following changes:

In order to create a new ticket in Zendesk, the code used to fetch ID of the group to which a ticket should be assigned, by querying Zendesk's API. The query involved searching by the group's name. However, group names are prone to change, and so the code needed to be changed in order to keep it in sync.

This addresses this by allowing us to set the group's ID as an environment variable, thereby eliminating the need to query using group's name.

The README has been updated to explain what the environment variables
refer to.

https://trello.com/c/EhzSC1vv/366-zendesk-migration-configure-rp-form